Project Update.
Fixed issue where all of the if statements in generateParagraphs were being executed on every iteration of the for loop, but they were mutually exclusive. This happened when I changed from using .every to using a for loop. Optimized our for loops and array 'stuff,' ie pre-allocating arrays. Set arrays to null when finished with them. Now have annoying bug where we have blank values and undefined at or near the end of our paragraphs array. Undoubtedly caused by pre-allocation.