Project Update.
Added inputValueAndGenerate to App.test.js to group actions required by a number of tests. Added a test to check for the success alert after the user inputs a value and hits the 'I want that thing' button.' This then caused an issue with the progress bar test. Running both tests together would cause the success alert test to fail. Fixed this with a pause, but now getting act warnings. Can possibly address these by updating inputValueAndGenerate to use findBy instead of getBy...