Project Update.
Added waitFor to our progress bar test so that we wait for our paragraphs to be rendered. This has eliminated our act warnings and our progress bar test is passing as is our success alert test (which also uses waitFor to get paragraphs before checking for the success alert message). This isn't ideal as it seems that there is a leak from our progress bar test to our success alert test, but at least they are passing when run together.